requirements.txt
influxdb-python>=0.1.7